Update seismic_service_datatypes.proto
We are missing description for name and LineRange in Seismic. Can you please add it when reviewing this PR. Thanks!

In Seismic Store. It says name will be change to external-id. Do we need to add an extra field or edit this one?

Is the description for storage_tier_name correct?