Add a GetVolumeBounds rpc for estimating size
When supporting geometry queries in the V1 API, we'll want to compute
the total size of returned traces (for `volume_seismic.get_volume_size`)
as well as the min, max, step encompassing all inlines and xlines (for
finding proper array sizes in `volume_seismic.get_array`). This is
doable from just the volume def with line-based queries, but not with
spatial queries. So instead of adding more and more code duplication to
the sdk, perform the computation on the server side with a new rpc.