Add endpoint to debug ingestion (#206)
* Proposal: Endpoint to debug ingestion

Some failed ingestion results are really hard to debug. It can help to
get a look at what the ingestion worker has seen so far before failing,
such as trace extents. Here is a suggested endpoint to peek at the
unpublished seismicstore resulting from a failed ingestion, in order to
debug.

Alternatively, we could add `allow_unpublished` to
`SearchSeismicStoresRequest`, but keeping it as a separate endpoint
which is explicitly meant for debugging seems to more clearly
communicate the debugging intent.

Are there any potential consistency problems with allowing data managers
to see unpublished seismic stores like this?

* Feedback